+++ Scripts/Items/Temrael/Armures/ArmureDrow.cs
@@ -19,8 +19,8 @@ namespace Server.Items
             }
         }
 
-        public override int BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorChainNoir.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorChainNoir.max_Durabilite; } }
@@ -72,8 +72,8 @@ namespace Server.Items
             }
         }
 
-        public override int BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorChainNoir.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorChainNoir.max_Durabilite; } }
@@ -125,8 +125,8 @@ namespace Server.Items
             }
         }
 
-        public override int BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorChainNoir.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorChainNoir.max_Durabilite; } }
@@ -178,8 +178,8 @@ namespace Server.Items
             }
         }
 
-        public override int BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorChainNoir.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorChainNoir.max_Durabilite; } }
@@ -231,8 +231,8 @@ namespace Server.Items
             }
         }
 
-        public override int BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorChainNoir.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorChainNoir.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorChainNoir.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorChainNoir.max_Durabilite; } }
