+++ Scripts/Items/Temrael/Armures/Bourgeon.cs
@@ -7,8 +7,8 @@ namespace Server.Items
     {
         //public override int NiveauAttirail { get { return Bourgeon_Niveau; } }
 
-        public override int BasePhysicalResistance { get { return ArmorBourgeon.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorBourgeon.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorBourgeon.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorBourgeon.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorBourgeon.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorBourgeon.max_Durabilite; } }
@@ -48,8 +48,8 @@ namespace Server.Items
     {
         //public override int NiveauAttirail { get { return Bourgeon_Niveau; } }
 
-        public override int BasePhysicalResistance { get { return ArmorBourgeon.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorBourgeon.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorBourgeon.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorBourgeon.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorBourgeon.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorBourgeon.max_Durabilite; } }
@@ -89,8 +89,8 @@ namespace Server.Items
     {
         //public override int NiveauAttirail { get { return Bourgeon_Niveau; } }
 
-        public override int BasePhysicalResistance { get { return ArmorBourgeon.resistance_Physique; } }
-        public override int BaseMagieResistance { get { return ArmorBourgeon.resistance_Magique; } }
+        public override double BasePhysicalResistance { get { return ArmorBourgeon.resistance_Physique; } }
+        public override double BaseMagieResistance { get { return ArmorBourgeon.resistance_Magique; } }
 
         public override int InitMinHits { get { return ArmorBourgeon.min_Durabilite; } }
         public override int InitMaxHits { get { return ArmorBourgeon.max_Durabilite; } }
